Create new branch to be used for the ROB10 project
Added a new branch where Alexandru Mihai Smau and Bogdan Stefan
Miron will add their files and source code.
This branch will expand on the previous team's work on the Mars
Rover. Speciffically, we will integarte ROS2 with isaac sim in
regards to the rover. On completion of the project, the owners of
the masters branch are free to cherry-pick whatever they want
(probably all, since Bogdan and I are awesome).