Added the rover controller and set up an action graph.
Populated the controller's Python script with the kinematic model.
Updated the Action Graph to connect the controller to the rover
model. Note that the rover requires steering angles and angular
velocities for the wheels. The articulation controller can only
do angles or angular velocities. As such, two articulation
controllers are needed.

Note: Currently, the rover cannot turn on its center. This is the
subject of followup work.