Try the carter code again with single robot.
Copy all assets needed for the environment and save it locally.
Set the appropriate stereo camera values.
Set all ros contexts to 5. This is done in all 3 action graphs.

I suspect that the /tf does not connect all the frames correctly,
because the base is spinning, but the camera is not....