Added the IMU sensor
Added IMU sendor and published the data as a ROS2 topic, and it
actually works, holy crap.

NOTE: The legacy IMU sensor extension has to be disabled before
loading the environment.

NOTE: Make sure that you do not chown the isaac sim fodler before
OR after starting the container. Make sure you are on a clean
branch then reboot. Really stupid, but don't lnow of a better way.