Adding rover and maps and cameras
Combine rover usda from AAU with map1 from Luxemburg. Adding cameras
to the rover such a stereo camera is simmulated. Created also
action graphs for the cameras to communicate with ros2.

When the simmulation is runned the rover is falling to test the
gravity part and 2 viewports are "auto-launched" showing that the
output of the camera is viable